Sliders working

+++ app/src/main/java/com/stevedenheyer/scriptassistant/common/domain/model/audio/Settings.kt
@@ -1,6 +1,6 @@
 package com.stevedenheyer.scriptassistant.common.domain.model.audio
 
 data class Settings(
-    val threshold: Int,
-    val pause: Int
+    val threshold: Float,
+    val pause: Float
 )
